+++ labs/00-Programming/src/18.c
@@ -8,11 +8,11 @@
 #define MAX 10000
 
 
-int transpose_matrix(int** matrix, int height);
+int transpose_matrix(double** matrix, int height);
 
 
 int program_00_18() {
-    int** matrix = random_matrix_generate(MATRIX_SIZE, MATRIX_SIZE, 100, -100);
+    double** matrix = random_matrix_generate(MATRIX_SIZE, MATRIX_SIZE, 100, -100);
 
     if (matrix == NULL) {
         printf("Memory alocation error. Exit code 1.");
@@ -39,7 +39,7 @@ int program_00_18() {
 }
 
 
-int transpose_matrix(int** matrix, int size) {
+int transpose_matrix(double** matrix, int size) {
     for (int i = 0; i < size; i++) {
         for (int j = i + 1; j < size; j++) {
             int temp = matrix[i][j];
